change to BigDecimal calculations

+++ app/src/main/java/com/aintogarov/currencyconverter/domain/CurrencyAmount.kt
@@ -1,7 +1,9 @@
 package com.aintogarov.currencyconverter.domain
 
+import java.math.BigDecimal
+
 
 data class CurrencyAmount(
     val currency: String,
-    val value: Double
+    val value: BigDecimal
 )
